q1/requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
joblib>=1.4
pyyaml>=6.0
matplotlib>=3.9
seaborn>=0.13
openpyxl>=3.1.0
